Move ProxyPromiseImpl
